Fixed clippy warnings

+++ ipp/src/value.rs
@@ -418,7 +418,7 @@ mod tests {
     fn value_check(value: IppValue) {
         let mut b = value.to_bytes();
         b.advance(2); // skip value size
-        assert_eq!(IppValue::parse(value.to_tag() as u8, b).unwrap(), value);
+        assert_eq!(IppValue::parse(value.to_tag(), b).unwrap(), value);
     }
 
     #[test]
